Did quite a lot on the teacher part. Actually setup a lot of other utils and funcs. Tomorrow continuing work, possibly and hopefully finishing the api.

+++ models/TermMarkModel.go
@@ -1,9 +1,9 @@
 package models
 
 type TermMark struct {
-	TermMarkID string `json:"termMarkID,omitempty" bson:"termMarkID,omitempty"`
-	Value      int    `json:"value,omitempty" bson:"value,omitempty"`
-	StudentID  string `json:"studentID,omitempty" bson:"studentID,omitempty"`
-	Grade      Grade  `json:"grade,omitempty" bson:"grade,omitempty"`
-	Term       int    `json:"term,omitempty" bson:"term,omitempty"`
+	TermMarkID string  `json:"termMarkID,omitempty" bson:"termMarkID,omitempty"`
+	Value      float64 `json:"value,omitempty" bson:"value,omitempty"`
+	StudentID  string  `json:"studentID,omitempty" bson:"studentID,omitempty"`
+	Grade      Grade   `json:"grade,omitempty" bson:"grade,omitempty"`
+	Term       int     `json:"term,omitempty" bson:"term,omitempty"`
 }
